# [next] Correct output file tracing and limit calculation