--- a/packages/next/src/server-build.ts
+++ b/packages/next/src/server-build.ts
@@ -303,7 +303,6 @@ export async function serverBuild({
 
   if (hasLambdas) {
     const initialTracingLabel = 'Traced Next.js server files in';
-
     console.time(initialTracingLabel);
 
     const initialTracedFiles: {
@@ -412,19 +411,24 @@ export async function serverBuild({
 
     debug('collecting initial Next.js server files');
     await Promise.all(
-      initialFileList.map(
-        collectTracedFiles(
+      initialFileList.map(async file => {
+        await collectTracedFiles(
+          file,
           baseDir,
           lstatResults,
           lstatSema,
           initialFileReasons,
           initialTracedFiles
-        )
-      )
+        );
+      })
     );
 
     debug('creating initial pseudo layer');
-    const initialPseudoLayer = await createPseudoLayer(initialTracedFiles);
+    const initialPseudoLayer = await createPseudoLayer({
+      files: initialTracedFiles,
+      lstatResults,
+      lstatSema,
+    });
     console.timeEnd(initialTracingLabel);
 
     const lambdaCreationLabel = 'Created all serverless functions in';
@@ -435,18 +439,18 @@ export async function serverBuild({
     const appRouterPages: string[] = [];
     const appRouteHandlers: string[] = [];
 
-    lambdaPageKeys.forEach(page => {
+    for (const page of lambdaPageKeys) {
       if (
         internalPages.includes(page) &&
         page !== '404.js' &&
         !(page === '_error.js' && !(static404Page || lambdaPages['404.js']))
       ) {
-        return;
+        continue;
       }
       const pathname = page.replace(/\.js$/, '');
 
       if (nonLambdaSsgPages.has(pathname)) {
-        return;
+        continue;
       }
       const normalizedPathname = normalizePage(pathname);
 
@@ -465,7 +469,7 @@ export async function serverBuild({
       } else {
         nonApiPages.push(page);
       }
-    });
+    }
 
     const requiredFiles: { [key: string]: FileFsRef } = {};
 
@@ -557,7 +561,11 @@ export async function serverBuild({
 
     // add required files and internal pages to initial pseudo layer
     // so that we account for these in the size of each page group
-    const requiredFilesLayer = await createPseudoLayer(requiredFiles);
+    const requiredFilesLayer = await createPseudoLayer({
+      files: requiredFiles,
+      lstatResults,
+      lstatSema,
+    });
     Object.assign(
       initialPseudoLayer.pseudoLayer,
       requiredFilesLayer.pseudoLayer
@@ -681,16 +689,15 @@ export async function serverBuild({
       return originalPagePath;
     };
 
-    const getBuildTraceFile = (page: string) => {
-      return (
-        pageBuildTraces[page + '.nft.json'] ||
-        appBuildTraces[page + '.nft.json']
-      );
-    };
+    const getBuildTraceFile = (page: string) =>
+      pageBuildTraces[page + '.nft.json'] || appBuildTraces[page + '.nft.json'];
 
     const pathsToTrace: string[] = mergedPageKeys
       .map(page => {
-        if (!getBuildTraceFile(page)) {
+        const originalPagePath = getOriginalPagePath(page);
+        const pageBuildTrace = getBuildTraceFile(originalPagePath);
+
+        if (!pageBuildTrace) {
           return lambdaPages[page].fsPath;
         }
       })
@@ -700,6 +707,10 @@ export async function serverBuild({
     let parentFilesMap: Map<string, Set<string>> | undefined;
 
     if (pathsToTrace.length > 0) {
+      console.warn(
+        `Warning: running tracing as trace files were not present for:\n`,
+        JSON.stringify(pathsToTrace, null, 2)
+      );
       traceResult = await nodeFileTrace(pathsToTrace, {
         base: baseDir,
         cache: traceCache,
@@ -783,31 +794,39 @@ export async function serverBuild({
         reasons = traceResult?.reasons || new Map();
       }
 
-      await Promise.all(
-        fileList.map(
-          collectTracedFiles(
+      await Promise.all([
+        createPseudoLayer({
+          files: {
+            [page]: lambdaPages[page],
+          },
+          lstatResults,
+          lstatSema,
+        }).then(res => {
+          compressedPages[page] = res.pseudoLayer[page] as PseudoFile;
+        }),
+        ...fileList.map(async file => {
+          await collectTracedFiles(
+            file,
             baseDir,
             lstatResults,
             lstatSema,
             reasons,
             tracedFiles
-          )
-        )
-      );
+          );
+        }),
+      ]);
+
       pageTraces[page] = tracedFiles;
-      compressedPages[page] = (
-        await createPseudoLayer({
-          [page]: lambdaPages[page],
-        })
-      ).pseudoLayer[page] as PseudoFile;
     }
 
-    const tracedPseudoLayer = await createPseudoLayer(
-      mergedPageKeys.reduce((prev, page) => {
+    const tracedPseudoLayer = await createPseudoLayer({
+      files: mergedPageKeys.reduce((prev, page) => {
         Object.assign(prev, pageTraces[page]);
         return prev;
-      }, {})
-    );
+      }, {}),
+      lstatResults,
+      lstatSema,
+    });
 
     const pageExtensions = requiredServerFilesManifest.config?.pageExtensions;
 
@@ -946,144 +965,150 @@ export async function serverBuild({
       compressedPages
     );
 
-    for (const group of combinedGroups) {
-      const groupPageFiles: { [key: string]: PseudoFile } = {};
+    await Promise.all(
+      combinedGroups.map(async group => {
+        const groupPageFiles: { [key: string]: PseudoFile } = {};
 
-      for (const page of [...group.pages, ...internalPages]) {
-        const pageFileName = path.normalize(
-          path.relative(baseDir, lambdaPages[page].fsPath)
-        );
-        groupPageFiles[pageFileName] = compressedPages[page];
-      }
+        for (const page of [...group.pages, ...internalPages]) {
+          const pageFileName = path.normalize(
+            path.relative(baseDir, lambdaPages[page].fsPath)
+          );
+          groupPageFiles[pageFileName] = compressedPages[page];
+        }
 
-      const updatedManifestFiles: { [name: string]: FileBlob } = {};
+        const updatedManifestFiles: { [name: string]: FileBlob } = {};
 
-      if (isCorrectManifests) {
-        // filter dynamic routes to only the included dynamic routes
-        // in this specific serverless function so that we don't
-        // accidentally match a dynamic route while resolving that
-        // is not actually in this specific serverless function
-        for (const manifest of [
-          'routes-manifest.json',
-          'server/pages-manifest.json',
-        ] as const) {
-          const fsPath = path.join(entryPath, outputDirectory, manifest);
+        if (isCorrectManifests) {
+          // filter dynamic routes to only the included dynamic routes
+          // in this specific serverless function so that we don't
+          // accidentally match a dynamic route while resolving that
+          // is not actually in this specific serverless function
+          for (const manifest of [
+            'routes-manifest.json',
+            'server/pages-manifest.json',
+          ] as const) {
+            const fsPath = path.join(entryPath, outputDirectory, manifest);
 
-          const relativePath = path.relative(baseDir, fsPath);
-          delete group.pseudoLayer[relativePath];
+            const relativePath = path.relative(baseDir, fsPath);
+            delete group.pseudoLayer[relativePath];
 
-          const manifestData = await fs.readJSON(fsPath);
-          const normalizedPages = new Set(
-            group.pages.map(page => {
-              page = `/${page.replace(/\.js$/, '')}`;
-              if (page === '/index') page = '/';
-              return page;
-            })
+            const manifestData = await fs.readJSON(fsPath);
+            const normalizedPages = new Set(
+              group.pages.map(page => {
+                page = `/${page.replace(/\.js$/, '')}`;
+                if (page === '/index') page = '/';
+                return page;
+              })
+            );
+
+            switch (manifest) {
+              case 'routes-manifest.json': {
+                const filterItem = (item: { page: string }) =>
+                  normalizedPages.has(item.page);
+
+                manifestData.dynamicRoutes =
+                  manifestData.dynamicRoutes?.filter(filterItem);
+                manifestData.staticRoutes =
+                  manifestData.staticRoutes?.filter(filterItem);
+                break;
+              }
+              case 'server/pages-manifest.json': {
+                for (const key of Object.keys(manifestData)) {
+                  if (isDynamicRoute(key) && !normalizedPages.has(key)) {
+                    delete manifestData[key];
+                  }
+                }
+                break;
+              }
+              default: {
+                throw new NowBuildError({
+                  message: `Unexpected manifest value ${manifest}, please contact support if this continues`,
+                  code: 'NEXT_MANIFEST_INVARIANT',
+                });
+              }
+            }
+
+            updatedManifestFiles[relativePath] = new FileBlob({
+              contentType: 'application/json',
+              data: JSON.stringify(manifestData),
+            });
+          }
+        }
+
+        const launcherFiles: { [name: string]: FileFsRef | FileBlob } = {
+          [path.join(
+            path.relative(baseDir, projectDir),
+            '___next_launcher.cjs'
+          )]: new FileBlob({
+            data: group.isAppRouter ? appLauncher : launcher,
+          }),
+        };
+        const operationType = getOperationType({ group, prerenderManifest });
+
+        const lambda = await createLambdaFromPseudoLayers({
+          files: {
+            ...launcherFiles,
+            ...updatedManifestFiles,
+          },
+          layers: [group.pseudoLayer, groupPageFiles],
+          handler: path.join(
+            path.relative(baseDir, projectDir),
+            '___next_launcher.cjs'
+          ),
+          operationType,
+          memory: group.memory,
+          runtime: nodeVersion.runtime,
+          maxDuration: group.maxDuration,
+          isStreaming: group.isStreaming,
+          nextVersion,
+        });
+
+        for (const page of group.pages) {
+          const pageNoExt = page.replace(/\.js$/, '');
+          let isPrerender = prerenderRoutes.has(
+            path.join('/', pageNoExt === 'index' ? '' : pageNoExt)
           );
 
-          switch (manifest) {
-            case 'routes-manifest.json': {
-              const filterItem = (item: { page: string }) =>
-                normalizedPages.has(item.page);
-
-              manifestData.dynamicRoutes =
-                manifestData.dynamicRoutes?.filter(filterItem);
-              manifestData.staticRoutes =
-                manifestData.staticRoutes?.filter(filterItem);
-              break;
-            }
-            case 'server/pages-manifest.json': {
-              for (const key of Object.keys(manifestData)) {
-                if (isDynamicRoute(key) && !normalizedPages.has(key)) {
-                  delete manifestData[key];
-                }
-              }
-              break;
-            }
-            default: {
-              throw new NowBuildError({
-                message: `Unexpected manifest value ${manifest}, please contact support if this continues`,
-                code: 'NEXT_MANIFEST_INVARIANT',
-              });
-            }
+          if (!isPrerender && routesManifest?.i18n) {
+            isPrerender = routesManifest.i18n.locales.some(locale => {
+              return prerenderRoutes.has(
+                path.join('/', locale, pageNoExt === 'index' ? '' : pageNoExt)
+              );
+            });
           }
 
-          updatedManifestFiles[relativePath] = new FileBlob({
-            contentType: 'application/json',
-            data: JSON.stringify(manifestData),
-          });
-        }
-      }
+          const outputName = normalizeIndexOutput(
+            path.posix.join(entryDirectory, pageNoExt),
+            true
+          );
 
-      const launcherFiles: { [name: string]: FileFsRef | FileBlob } = {
-        [path.join(path.relative(baseDir, projectDir), '___next_launcher.cjs')]:
-          new FileBlob({ data: group.isAppRouter ? appLauncher : launcher }),
-      };
-      const operationType = getOperationType({ group, prerenderManifest });
-
-      const lambda = await createLambdaFromPseudoLayers({
-        files: {
-          ...launcherFiles,
-          ...updatedManifestFiles,
-        },
-        layers: [group.pseudoLayer, groupPageFiles],
-        handler: path.join(
-          path.relative(baseDir, projectDir),
-          '___next_launcher.cjs'
-        ),
-        operationType,
-        memory: group.memory,
-        runtime: nodeVersion.runtime,
-        maxDuration: group.maxDuration,
-        isStreaming: group.isStreaming,
-        nextVersion,
-      });
-
-      for (const page of group.pages) {
-        const pageNoExt = page.replace(/\.js$/, '');
-        let isPrerender = prerenderRoutes.has(
-          path.join('/', pageNoExt === 'index' ? '' : pageNoExt)
-        );
-
-        if (!isPrerender && routesManifest?.i18n) {
-          isPrerender = routesManifest.i18n.locales.some(locale => {
-            return prerenderRoutes.has(
-              path.join('/', locale, pageNoExt === 'index' ? '' : pageNoExt)
-            );
-          });
-        }
-
-        const outputName = normalizeIndexOutput(
-          path.posix.join(entryDirectory, pageNoExt),
-          true
-        );
-
-        // we add locale prefixed outputs for SSR pages,
-        // this is handled in onPrerenderRoute for SSG pages
-        if (
-          i18n &&
-          !isPrerender &&
-          !group.isAppRouter &&
-          (!isCorrectLocaleAPIRoutes ||
-            !(pageNoExt === 'api' || pageNoExt.startsWith('api/')))
-        ) {
-          for (const locale of i18n.locales) {
-            lambdas[
-              normalizeIndexOutput(
-                path.posix.join(
-                  entryDirectory,
-                  locale,
-                  pageNoExt === 'index' ? '' : pageNoExt
-                ),
-                true
-              )
-            ] = lambda;
+          // we add locale prefixed outputs for SSR pages,
+          // this is handled in onPrerenderRoute for SSG pages
+          if (
+            i18n &&
+            !isPrerender &&
+            !group.isAppRouter &&
+            (!isCorrectLocaleAPIRoutes ||
+              !(pageNoExt === 'api' || pageNoExt.startsWith('api/')))
+          ) {
+            for (const locale of i18n.locales) {
+              lambdas[
+                normalizeIndexOutput(
+                  path.posix.join(
+                    entryDirectory,
+                    locale,
+                    pageNoExt === 'index' ? '' : pageNoExt
+                  ),
+                  true
+                )
+              ] = lambda;
+            }
+          } else {
+            lambdas[outputName] = lambda;
           }
-        } else {
-          lambdas[outputName] = lambda;
         }
-      }
-    }
+      })
+    );
     console.timeEnd(lambdaCreationLabel);
   }