--- a/packages/next/src/utils.ts
+++ b/packages/next/src/utils.ts
@@ -2521,10 +2521,6 @@ export function normalizeEdgeFunctionPath(
     shortPath = 'index';
   }
 
-  if (shortPath.startsWith('pages/')) {
-    shortPath = shortPath.replace(/^pages\//, '');
-  }
-
   return shortPath;
 }