--- a/packages/next/test/unit/fixtures/01-normalize-paths/index.test.js
+++ b/packages/next/test/unit/fixtures/01-normalize-paths/index.test.js
@@ -24,7 +24,9 @@ describe(`${__dirname.split(path.sep).pop()}`, () => {
     }, {});
 
     const { output, routes } = await build({
-      config: {},
+      config: {
+        installCommand: 'yarn',
+      },
       entrypoint: 'package.json',
       files,
       meta: {