… .changeset/bright-dolls-drop.md
… packages/next/src/constants.ts
… packages/next/test/integration/integration-2.test.js
… packages/next/test/unit/fixtures/01-normalize-paths/yarn.lock
… packages/next/test/unit/fixtures/02-edge-function-basepath/yarn.lock
… packages/next/test/unit/fixtures/03-with-api-routes/yarn.lock
… packages/next/test/unit/fixtures/04-edge-function-i18n/index.test.js